ir_nec: improve robustness, present more data on invalid input
Don't terminate IR frame inspection when successfully decoded fields
fail the validity check. Reset internal state on long periods of idle
level instead (beware of the late detection of the STOP condition).

Unconditionally annotate received fields, and optionally amend them with
a warning annotation. It's unexpected to not see the address field value
at all just because the number was not acceptable for the potentially
incorrect protocol selection.

Rename a variable in the decode() routine. The 'b' identifier suggested
a bit value, but it was used for the pulse width.