sle44xx: fixup boilerplate, add missing required fields
The decoder was created in an older setup and would not load in recent
applications. Add missing boilerplate which is mandatory. Rename 'data'
since annotation classes and rows must not have ambiguous names, and
there is no singular form of 'data' -- use 'fields' instead. This also
fixes a missing line termination at the end of the text file.