adf435x: Move decoder logic to SPI transfers
Previous decoder implementation was using a hardcoded falling CS
edge for triggering the decoding. This was the wrong edge per the
datasheet, where data is said to be shifted in on rising edge.

We now use the underlaying CS polarity from the SPI decoder by
triggering on the TRANSFER data type. As the adf435x decoder still
needs the CS polarity to be set correctly in the SPI decoder to
receive any data at all this makes the CS handling simpler and
less error prone.

This fixes bug https://sigrok.org/bugzilla/show_bug.cgi?id=1814

Reported-by:  Martin Homuth-Rosemann <homuth-rosemann@gmx.net>